+++ src/score/pub/types.ts
@@ -290,17 +290,6 @@ export type LyricsOptions = {
     hyphen?: LyricsHyphenValue;
 }
 
-/** @deprecated Fermata is deprecated. Will be removed in future release. */
-export enum Fermata {
-    /** Anchor fermata to note/rest. @deprecated */
-    AtNote = "atNote",
-    /** Anchor fermata to measure end. @deprecated */
-    AtMeasureEnd = "atMeasureEnd"
-}
-
-/** @deprecated */
-export type FermataValue = Fermata | `${Fermata}`;
-
 /** @deprecated Label is deprecated. Will be removed in future release. */
 export enum Label {
     /** "C", "C#", "Db", "D", etc. @deprecated */